Collapse utils module into single util.py file
- Move util.py from src/molecule/utils/ to src/molecule/
- Consolidate to_bool function from boolean.py into util.py
- Update all imports from molecule.utils.util to molecule.util
- Remove empty utils directory and associated files
- All existing functionality preserved
