+++ src/molecule/command/init/scenario.py
@@ -30,7 +30,7 @@ from typing import TYPE_CHECKING
 
 import click
 
-from molecule import logger
+from molecule import logger, util
 from molecule.click_cfg import click_command_ex
 from molecule.command.init import base
 from molecule.config import (
@@ -43,7 +43,6 @@ from molecule.constants import (
     MOLECULE_DEFAULT_SCENARIO_NAME,
     MOLECULE_ROOT,
 )
-from molecule.utils import util
 
 
 if TYPE_CHECKING:
