+++ src/molecule/dependency/base.py
@@ -28,8 +28,7 @@ import time
 from subprocess import CalledProcessError
 from typing import TYPE_CHECKING
 
-from molecule import logger
-from molecule.utils import util
+from molecule import logger, util
 
 
 if TYPE_CHECKING:
