+++ src/molecule/scenarios.py
@@ -25,9 +25,8 @@ import logging
 
 from typing import TYPE_CHECKING
 
-from molecule import ansi_output
+from molecule import ansi_output, util
 from molecule.reporting.definitions import ScenariosResults
-from molecule.utils import util
 
 
 if TYPE_CHECKING:
