--- src/molecule/utils/__init__.py
@@ -1,3 +1,0 @@
-"""Utilities package for molecule."""
-
-from __future__ import annotations
