+++ src/molecule/verifier/ansible.py
@@ -25,10 +25,9 @@ import os
 
 from typing import TYPE_CHECKING, cast
 
-from molecule import logger
+from molecule import logger, util
 from molecule.api import Verifier
 from molecule.reporting.definitions import CompletionState
-from molecule.utils import util
 
 
 if TYPE_CHECKING:
