+++ tests/unit/command/init/test_scenario.py
@@ -25,9 +25,9 @@ from typing import TYPE_CHECKING
 
 import pytest
 
+from molecule import util
 from molecule.command.init.scenario import Scenario
 from molecule.config import Config
-from molecule.utils import util
 
 
 if TYPE_CHECKING:
