+++ tests/unit/command/test_side_effect.py
@@ -26,8 +26,8 @@ from typing import TYPE_CHECKING
 
 import pytest
 
+from molecule import util
 from molecule.command import side_effect
-from molecule.utils import util
 
 
 if TYPE_CHECKING:
