+++ tests/unit/provisioner/test_ansible_playbooks.py
@@ -25,8 +25,8 @@ from typing import TYPE_CHECKING
 
 import pytest
 
+from molecule import util
 from molecule.provisioner import ansible_playbooks
-from molecule.utils import util
 from tests.unit.conftest import os_split  # pylint:disable=C0411
