+++ tests/unit/test_config.py
@@ -27,10 +27,9 @@ from typing import TYPE_CHECKING, Any, Literal
 
 import pytest
 
-from molecule import config, platforms, scenario, state
+from molecule import config, platforms, scenario, state, util
 from molecule.dependency import ansible_galaxy, shell
 from molecule.provisioner import ansible
-from molecule.utils import util
 from molecule.verifier.ansible import Ansible as AnsibleVerifier
