+++ tests/unit/test_scenario.py
@@ -27,9 +27,8 @@ from typing import TYPE_CHECKING
 
 import pytest
 
-from molecule import config
+from molecule import config, util
 from molecule.scenario import Scenario
-from molecule.utils import util
 
 
 if TYPE_CHECKING:
