+++ tests/unit/test_util.py
@@ -30,9 +30,9 @@ from typing import TYPE_CHECKING
 import pytest
 import yaml
 
+from molecule import util
 from molecule.constants import MOLECULE_COLLECTION_GLOB, MOLECULE_HEADER
 from molecule.exceptions import MoleculeError
-from molecule.utils import util
 
 
 if TYPE_CHECKING:
@@ -628,7 +628,7 @@ def test_lookup_config_file_found_in_vcs_root(
     config_file.write_text("config: vcs_root")
 
     # Mock find_vcs_root to return our test directory
-    monkeypatch.setattr("molecule.utils.util.find_vcs_root", lambda default=None: str(vcs_root))  # noqa: ARG005
+    monkeypatch.setattr("molecule.util.find_vcs_root", lambda default=None: str(vcs_root))  # noqa: ARG005
 
     # Change to a subdirectory
     subdir = vcs_root / "subdir"
@@ -675,7 +675,7 @@ def test_lookup_config_file_found_in_collection_extensions(
     monkeypatch.chdir(collection_root)
 
     # Mock find_vcs_root to return empty (no VCS root)
-    monkeypatch.setattr("molecule.utils.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
+    monkeypatch.setattr("molecule.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
 
     # Clear caches
     util.get_collection_metadata.cache_clear()
@@ -707,10 +707,10 @@ def test_lookup_config_file_found_in_home_directory(
     config_file.write_text("config: home")
 
     # Mock Path.home() to return our fake home
-    monkeypatch.setattr("molecule.utils.util.Path.home", lambda: fake_home)
+    monkeypatch.setattr("molecule.util.Path.home", lambda: fake_home)
 
     # Mock find_vcs_root to return empty (no VCS root)
-    monkeypatch.setattr("molecule.utils.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
+    monkeypatch.setattr("molecule.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
 
     # Change to a different directory (not a collection)
     work_dir = tmp_path / "workdir"
@@ -763,8 +763,8 @@ def test_lookup_config_file_search_priority_vcs_root_wins(
     collection_config.write_text("config: collection")
 
     # Set up mocks
-    monkeypatch.setattr("molecule.utils.util.find_vcs_root", lambda default=None: str(vcs_root))  # noqa: ARG005
-    monkeypatch.setattr("molecule.utils.util.Path.home", lambda: fake_home)
+    monkeypatch.setattr("molecule.util.find_vcs_root", lambda default=None: str(vcs_root))  # noqa: ARG005
+    monkeypatch.setattr("molecule.util.Path.home", lambda: fake_home)
 
     # Change to collection directory
     monkeypatch.chdir(collection_root)
@@ -809,8 +809,8 @@ def test_lookup_config_file_search_priority_collection_over_home(
     collection_config.write_text("config: collection")
 
     # Set up mocks - no VCS root found
-    monkeypatch.setattr("molecule.utils.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
-    monkeypatch.setattr("molecule.utils.util.Path.home", lambda: fake_home)
+    monkeypatch.setattr("molecule.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
+    monkeypatch.setattr("molecule.util.Path.home", lambda: fake_home)
 
     # Change to collection directory
     monkeypatch.chdir(collection_root)
@@ -842,8 +842,8 @@ def test_lookup_config_file_not_found(
     work_dir.mkdir()
 
     # Set up mocks - no VCS root, no collection
-    monkeypatch.setattr("molecule.utils.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
-    monkeypatch.setattr("molecule.utils.util.Path.home", lambda: fake_home)
+    monkeypatch.setattr("molecule.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
+    monkeypatch.setattr("molecule.util.Path.home", lambda: fake_home)
 
     # Change to work directory
     monkeypatch.chdir(work_dir)
@@ -887,7 +887,7 @@ def test_lookup_config_file_handles_path_filename_extraction(
     monkeypatch.chdir(collection_root)
 
     # Mock find_vcs_root to return empty
-    monkeypatch.setattr("molecule.utils.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
+    monkeypatch.setattr("molecule.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
 
     # Clear caches
     util.get_collection_metadata.cache_clear()
@@ -925,8 +925,8 @@ def test_lookup_config_file_collection_no_extensions_dir(
     home_config.write_text("config: home")
 
     # Set up mocks
-    monkeypatch.setattr("molecule.utils.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
-    monkeypatch.setattr("molecule.utils.util.Path.home", lambda: fake_home)
+    monkeypatch.setattr("molecule.util.find_vcs_root", lambda default=None: "~")  # noqa: ARG005
+    monkeypatch.setattr("molecule.util.Path.home", lambda: fake_home)
 
     # Change to collection directory
     monkeypatch.chdir(collection_root)
