+++ tests/unit/verifier/test_testinfra.py
@@ -24,7 +24,7 @@ import os
 import pytest
 
 from molecule.config import Config
-from molecule.utils.util import write_file
+from molecule.util import write_file
 from molecule.verifier import testinfra
